refactor(ui): extract blob download helper
